feat: consolidate component tests
